Adjust Model to account for the fact that knex might not be available during construction
Methods that require the knex instance will throw an Error if it's not available yet, instructing the user to connect before making any queries. It's the simplest approach, but could be replaced by some sort of queuing, where the query will be executed as soon as a connection becomes available. However, knex doesn't really seem to give you this kind of control: creating a knex instance seems to automatically start connection (afaik).